+++ R/make_trial_dataset.R
@@ -6,7 +6,7 @@ make_trial_dataset = function() {
   n <- 100 # Number of patients
   pr <- 0.8 # Proportion of resistant patients
   pe <- 0.2 # Placebo effect on biomarker levels
-  de <- 0.5 # Drug effect on biomarker levels
+  de <- 0.3 # Drug effect on biomarker levels
   sd <- 0.05 # Patient and drug effect individual differences (noise)
   sickbiom <- 0.3 # Mean of biomarker level for sick patients
   healthybiom <- 0.7 # Value above which patients are considered healthy
@@ -20,9 +20,11 @@ make_trial_dataset = function() {
   # Randomly allocated unhealthy biomarker levels to patients: pretreatment biomarker levels
   # Simulated with a normal distribution centred around 0.3 with a standard deviation of 0.1
   prebiom <- stats::rnorm(n, mean = sickbiom, sd = sd)
+  prebiom <- pmax(pmin(prebiom, 1), 0)
 
   # Simulate treatment of patients according to their resistance, drug intake and pretreatment biomarker levels
-  postbiom <- prebiom + pe  * (1 - (drug == "drug")) + de * (resistance == "sensitive") * (drug == "drug") + stats::rnorm(n, mean = 0, sd = sd)
+  postbiom <- prebiom + pe + de * (resistance == "sensitive") * (drug == "drug") + stats::rnorm(n, mean = 0, sd = sd)
+  postbiom <- pmax(pmin(postbiom, 1), 0)
 
   # Summarize results in data.frame and save
   trial <- data.frame(drug, resistance, prebiom, postbiom)